requirements.txt
asgiref==3.8.1
attrs==24.3.0
boto3==1.35.71
botocore==1.35.71
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-request==1.6.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
